debug logging

+++ mouse.sh
@@ -11,23 +11,25 @@ res=`nc -v -l -q 0 $port`
 while true; do
 
 	echo "MOUSE: STARTING LOOP" >> ex2.log
-	
-	if [ ${res:${#res} - 4} == "MEOW" ]
+	if [ -n $res ]
 	then
-		echo "I WAS CAUGHT"
-		name=${res:0:5}
-		echo "$name"
-		if [ $name == "Catty" ]
+		if [ ${res:${#res} - 4} == "MEOW" ]
 		then
-			pid=$(cat "cattypid")
-			kill -INT $pid
-		elif [ $name == "Jazzy" ]
-		then
-			pid=$(cat "jazzypid")
-			kill -INT $pid
+			echo "I WAS CAUGHT"
+			name=${res:0:5}
+			echo "$name"
+			if [ $name == "Catty" ]
+			then
+				pid=$(cat "cattypid")
+				kill -INT $pid
+			elif [ $name == "Jazzy" ]
+			then
+				pid=$(cat "jazzypid")
+				kill -INT $pid
+			fi
+			break
+		else
+			res=`nc -v -l -q 0 $port`
 		fi
-		break
-	else
-		res=`nc -v -l -q 0 $port`
 	fi
 done
